merge from original repo

+++ Snap_rounding_2/include/CGAL/Snap_rounding_2.h
@@ -20,6 +20,9 @@
 #ifndef CGAL_SNAP_ROUNDING_2_H
 #define CGAL_SNAP_ROUNDING_2_H
 
+#include <CGAL/license/Snap_rounding_2.h>
+
+
 #include <iostream>
 #include <CGAL/basic.h>
 #include <CGAL/enum.h>
@@ -614,7 +617,7 @@ find_intersected_hot_pixels(Segment_data & seg,
 
   }
 
-  number_of_intersections = hot_pixels_intersected_set.size();
+  number_of_intersections = static_cast<int>(hot_pixels_intersected_set.size());
 }
 
 /*! */
